XRDDEV-323 remove demo code: City repository and controller
