+++ src/proxy-ui-api/frontend/vue.config.js
@@ -1,6 +1,6 @@
 module.exports = {
   devServer: {
-    proxy: 'http://xroad2-docker-ss6:8020' // this needs to be parametrized (address where backend runs)
+    proxy: 'http://xroad2-docker-ss6.new.local:8020' // this needs to be parametrized (address where backend runs)
   },
 
   pluginOptions: {
